+++ core/src/lib.rs
@@ -35,6 +35,9 @@
 //!   define how to upgrade each individual substream to use a protocol.
 //!   See the `upgrade` module.
 
+#[cfg(feature = "serde")]
+extern crate _serde as serde;
+
 mod keys_proto {
     include!(concat!(env!("OUT_DIR"), "/keys_proto.rs"));
 }