+++ protocols/gossipsub/CHANGELOG.md
@@ -9,11 +9,14 @@
 - Improve bandwidth performance by tracking IWANTs and reducing duplicate sends
   (see [PR 2327]).
 
+- Implement `Serialize` and `Deserialize` for `MessageId` and `FastMessageId` (see [PR 2408])
+
 - Fix `GossipsubConfigBuilder::build()` requiring `&self` to live for `'static` (see [PR 2409])
 
 [PR 2346]: https://github.com/libp2p/rust-libp2p/pull/2346
 [PR 2339]: https://github.com/libp2p/rust-libp2p/pull/2339
 [PR 2327]: https://github.com/libp2p/rust-libp2p/pull/2327
+[PR 2408]: https://github.com/libp2p/rust-libp2p/pull/2408
 [PR 2409]: https://github.com/libp2p/rust-libp2p/pull/2409
 
 # 0.34.0 [2021-11-16]